css/css-contain/container-queries/top-layer-dialog-container.html

<!doctype html>
<title>CSS Container Queries Test: Top layer element as a @container</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-queries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #parent { width: 100px; }
  #dialog {
    container-type: inline-size;
    width: auto;
    border: none;
  }
  #child { color: red; }
  @container (min-width: 200px) {
    #child { color: green; }
  }
</style>
<div id="parent">
  <dialog id="dialog"><span id="child"></span></dialog>
</div>
<script>
  setup(() => assert_implements_container_queries());

  promise_test(async () => {
    if (waitForPolyfill) await waitForPolyfill();
    assert_equals(getComputedStyle(child).color, "rgb(255, 0, 0)");
  }, "#dialog initially sized by #containing-block");

  promise_test(async (t) => {
    assert_implements_optional(typeof dialog.showModal === 'function');
    t.add_cleanup(() => {
      dialog.close();
    });
    dialog.showModal();
    if (waitForPolyfill) await waitForPolyfill();
    assert_equals(getComputedStyle(child).color, "rgb(0, 128, 0)");
  }, "#dialog sized by viewport");
</script>
